Fix compilation bugs

+++ fasteq/ops/equi_linear.py
@@ -38,13 +38,13 @@ class _FastEquiLinearFn(torch.autograd.Function):
         #cg_vals = [0.03857583749052298, 0.02988071523335984, 0.03340765523905305]
         #I_total = sum(I_list)
 
-        x = x.float()
-        w = w.float()
+        # x = x.float()
+        # w = w.float()
 
         out = torch.ops.equi_linear.forward(x, w, I_list, cg_list)
         out =  out.view(B, -1)
 
-        out = out.double()
+        # out = out.double()
 
         #torch.cuda.synchronize()
         #end_time = time.perf_counter() * 1000
@@ -70,15 +70,15 @@ class _FastEquiLinearFn(torch.autograd.Function):
 
         print(f"grad_out shape:{grad_out.shape}, w shape:{w.shape}")
 
-        grad_out = grad_out.float()
-        w = w.float()
+        # grad_out = grad_out.float()
+        # w = w.float()
 
         grad_out = grad_out.view(ctx.B, 16, ctx.u).contiguous() # 只支持out固定为16
         grad_x = torch.ops.equi_linear.backward(grad_out, w, ctx.I_list, ctx.cg_list).view(ctx.B, -1)
 
         print(f"grad_x shape after bwd:{grad_x.shape}")
 
-        grad_x = grad_x.double()
+        # grad_x = grad_x.double()
 
         torch.cuda.synchronize()
         end_time = time.perf_counter() * 1000
